(squashed): move commits from datapane-next
