+++ webui/report/index.ts
@@ -1,5 +1,5 @@
 import { createApp, defineCustomElement } from "vue";
-import Report from "./src/components/Report.vue";
+import Report from "./src/components/ReportComponent.vue";
 import TableBlock from "./src/components/blocks/Table.ce.vue";
 import "./src/styles/report.scss";
 import "highlight.js/styles/stackoverflow-light.css";
